Define more modules as SPM targets
